+++ OLGA/OLGA-Core-CSharp-Generator/src/main/java/semanticstore/ontology/library/generator/generators/TrinityGenerator.java
