+++ OLGA/OLGA-Core-Java-Generator/src/main/java/semanticstore/ontology/library/generator/generators/RDF4JGenerator.java
@@ -54,12 +54,13 @@ import semanticstore.ontology.library.generator.model.ZClass;
 import semanticstore.ontology.library.generator.model.ZObjectProperty;
 import semanticstore.ontology.library.generator.resources.ResourceManager;
 
+@dagger.Module
 public class RDF4JGenerator extends AbstractGenerator {
 
   private String globalFile;
   private final static Logger log = Logger.getLogger(RDF4JGenerator.class);
 
-  public RDF4JGenerator() {
+  @Provides public RDF4JGenerator() {
     dataToBeInject = new HashMap<>();
     listOfParentZClass = new ArrayList<>();
   }
