+++ OLGA/OLGA-Core-Python-Generator/src/test/java/semanticstore/ontology/library/generator/code/generator/PreservePathOptionTestPython.java
@@ -34,39 +34,8 @@ public class PreservePathOptionTest {
 
   @BeforeClass
   public static void generateOntology() {
-    assertTrue(GeneratedOntologies.isSimpleBasicPreserveGenerated("cs", "trinity"));
-    assertTrue(GeneratedOntologies.isSimpleBasicGenerated("cs", "trinity"));
-    assertTrue(GeneratedOntologies.isSimpleBasicPreserveGenerated("java", "rdf4j"));
-    assertTrue(GeneratedOntologies.isSimpleBasicGenerated("java", "rdf4j"));
     assertTrue(GeneratedOntologies.isSimpleBasicPreserveGenerated("python", "rdfalchemy"));
     assertTrue(GeneratedOntologies.isSimpleBasicGenerated("python", "rdfalchemy"));
-  }
-
-  @Test
-  public void testPathWhenOptionSet_CSharp() {
-    File file = new File("OLGA/generated/SimplePreserve-dotnetTrinity/www/simple/com/Building.cs");
-    assertTrue(file.exists());
-  }
-
-  @Test
-  public void testPathWhenOptionNotSet_CSharp() {
-    File file =
-        new File("OLGA/generated/testSimple-dotnetTrinity/TestSimple/Rdf/Model/Building.cs");
-    assertTrue(file.exists());
-  }
-
-  @Test
-  public void testPathWhenOptionSet_Java() {
-    File file = new File(
-        "OLGA/generated/simplepreserve-RDF4J-java/src/main/java/www/simple/com/Building.java");
-    assertTrue(file.exists());
-  }
-
-  @Test
-  public void testPathWhenOptionNotSet_Java() {
-    File file = new File(
-        "OLGA/generated/testsimple-RDF4J-java/src/main/java/testsimple/rdf/model/Building.java");
-    assertTrue(file.exists());
   }
 
   @Test
