+++ OLGA/OLGA-Core/pom.xml
@@ -6,11 +6,11 @@
   <parent>
     <groupId>semanticstore.ontology.library.generator</groupId>
     <artifactId>OLGA</artifactId>
-    <version>0.0.5</version>
+    <version>0.0.6</version>
   </parent>  
   <artifactId>OLGA-Core</artifactId>
   <name>OLGA-Core</name>
-  <url>https://github.com/EcoStruxure/OLGA/tree/master/OLGA/OLGA-Core</url>
+	<url>https://github.com/EcoStruxure/OLGA/tree/master/OLGA/OLGA-Core</url>
   <description>OLGA is based on a model driven approach taking as input an ontology file expressed in one of the supported W3C supported standards (RDF, OWL) and generating a library conform to the ontology model.
 
 The generated library is then imported and used to programmatically to:
@@ -62,6 +62,13 @@ Query the generated ontology instance by relying on Object Oriented Classes.</de
 				<configuration>
 					<source>1.8</source>
 					<target>1.8</target>
+					<annotationProcessorPaths>
+						<path>
+							<groupId>com.google.dagger</groupId>
+							<artifactId>dagger-compiler</artifactId>
+							<version>2.28.3</version>
+						</path>
+					</annotationProcessorPaths>
 					<manifest>
 						<mainClass>semanticstore.ontology.library.generator.olga.OlgaService</mainClass>
 					</manifest>
